backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pymysql==1.1.1
paho-mqtt==2.1.0
scikit-learn==1.6.1
numpy==2.2.3
python-dotenv==1.0.1
